+++ cuvis/Calibration.py
@@ -48,6 +48,8 @@ class Calibration(object):
         return _id
 
     def __del__(self):
+        if self._handle is None:
+            return
         _ptr = cuvis_il.new_p_int()
         cuvis_il.p_int_assign(_ptr, self._handle)
         cuvis_il.cuvis_calib_free(_ptr)