+++ cuvis/SessionFile.py
@@ -108,6 +108,8 @@ class SessionFile(object):
         return self.get_measurement(key)
 
     def __del__(self):
+        if self._handle is None:
+            return
         _ptr = cuvis_il.new_p_int()
         cuvis_il.p_int_assign(_ptr, self._handle)
         cuvis_il.cuvis_session_file_free(_ptr)