+++ cuvis/Worker.py
@@ -332,6 +332,8 @@ class Worker(object):
 
     def __del__(self):
         self.reset_worker_callback()
+        if self._handle is None:
+            return
         _ptr = cuvis_il.new_p_int()
         cuvis_il.p_int_assign(_ptr, self._handle)
         cuvis_il.cuvis_worker_free(_ptr)